Replace TextLink with non react native TextLink
