+++ packages/link/README.md
@@ -1,7 +1,6 @@
 # Link
 
-The link component wraps its children and makes them clickable (web) or
-touchable (native) elements. There are two exported components:
+The link component wraps its children and makes them clickable web elements. There are two exported components:
 
 - `Link` - wraps any element, effectively making it clickable
 - `TextLink` - makes any inline text a link
@@ -22,13 +21,13 @@ import Link, { TextLink } from "@times-components/link";
 </TextLink>
 ```
 
-The web version of link also accepts a `responsiveLinkStyles` prop that allows
+The link accepts a `responsiveLinkStyles` prop that allows
 the resulting `<a>` tag to be styled responsively.
 
-Typically, apps will need to use the `onPress` callback to bubble up navigation
+<!-- Typically, apps will need to use the `onPress` callback to bubble up navigation
 so that the parent native Activity or UIViewContoller can handle the navigation,
 rather than multiple Activities or UIViewContollers mistakingly navigating
-multiple times in native land.
+multiple times in native land. -->
 
 ## Contributing
