+++ packages/pull-quote/__tests__/shared-with-style.js
@@ -15,7 +15,7 @@ jest.mock("@times-components/link", () => ({
   TextLink: "TextLink"
 }));
 jest.mock("@times-components/link", () => ({
-  TimesTextLink: "TimesTextLink"
+  TextLink: "TextLink"
 }));
 jest.mock("@times-components/icons", () => ({
   IconTwitter: "IconTwitter"
